Refactor the tests to make them passing on both platforms. testCriticalProcess() is stubbed
